volunteer details branch first commit, refactored users to volunteer to reflect consistent naming in folder directory
