+++ src/main/java/site/hhsa/demo/volunteers/controllers/VolunteerController.java
@@ -1,4 +1,4 @@
-package site.hhsa.demo.users.controllers;
+package site.hhsa.demo.volunteers.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import site.hhsa.demo.users.models.Volunteer;
-import site.hhsa.demo.users.repositories.VolunteerRepo;
+import site.hhsa.demo.volunteers.models.Volunteer;
+import site.hhsa.demo.volunteers.repositories.VolunteerRepo;
 
 import java.util.Date;
