+++ src/main/java/site/hhsa/demo/volunteers/models/VolunteerDetails.java
@@ -1,0 +1,4 @@
+package site.hhsa.demo.volunteers.models;
+
+public class VolunteerDetails {
+}
