+++ src/main/java/site/hhsa/demo/volunteers/repositories/VolunteerRepo.java
@@ -1,8 +1,8 @@
-package site.hhsa.demo.users.repositories;
+package site.hhsa.demo.volunteers.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import site.hhsa.demo.users.models.Volunteer;
+import site.hhsa.demo.volunteers.models.Volunteer;
 
 import java.util.List;
